修复个别类型错误

+++ src/main/java/com/example/house/dao/dos/Rule.java
@@ -54,8 +54,8 @@ public class Rule extends BaseEntity {
     private String creator;
 
     @TableField("version")
-    private long version;
+    private Integer version;
 
     @TableField("is_delete")
-    private long isDelete;
+    private Integer isDelete;
 }
